Shuffle teams within seeding classes
The team file is depicted so that the teams appear in the order in which
they are seeded. As many seeding classes are created as there are
groups.

So for example with four groups, the first four teams are first
seeds and will be seeded into different groups. Teams from the fifth team in
the file to eighth are second seeds and will be seeded into different
groups. And so on.