Use best game as tiebreaker in playoff seeding
Previously only the average score was used to determine the playoff
seeding. Now if the average score is even, use best game to determine
which of the teams is higher in the seeding.

If both the average score and best game of two or more advancing teams
is equal, the teams will be placed into the seedings in the order in
which they appeared in groups (i.e. team in group 1 is higher than a
team in group 2).

Since the group composition is also randomized, this should be a
relatively fair way to determine the seeding in the rare case that
multiple teams manage to advance with the same average score and best
game.