Fix wrong max group size being given to getGroups
The max group size is saved to file. This file is then read when an
existing schedule and group composition is being loaded.

The problem was that the max group size parameter should be given to
getGroups as is, but instead the calculated max group amount value was
passed to the function. The max group amount is calculated inside the
getGroups function, so in this case it was erroneusly calculated twice,
the first time receiving the correct value, and then making the
calculation on the correct value to receive some incorrect value.

As a result, too few groups were generated if the amount of teams was
high.